+++ Rules/IPv4/Outbound/Development/Microsoft/VisualStudio.ps1
@@ -340,6 +340,20 @@ experience improvement data uploader'" | Format-RuleOutput
 # Test if installation exists on system
 if ((Confirm-Installation "VisualStudioInstaller" ([ref] $VSInstallerRoot)) -or $ForceLoad)
 {
+	$Program = "$VSInstallerRoot\feedback.exe"
+	if ((Test-ExecutableFile $Program) -or $ForceLoad)
+	{
+		New-NetFirewallRule -DisplayName "VS Installer - feedback" `
+			-Platform $Platform -PolicyStore $PolicyStore -Profile $DefaultProfile `
+			-Service Any -Program $Program -Group $Group `
+			-Enabled True -Action Allow -Direction $Direction -Protocol TCP `
+			-LocalAddress Any -RemoteAddress Internet4 `
+			-LocalPort Any -RemotePort 443 `
+			-LocalUser $UsersGroupSDDL `
+			-InterfaceType $DefaultInterface `
+			-Description "Installer feedback" | Format-RuleOutput
+	}
+
 	$Program = "$VSInstallerRoot\resources\app\ServiceHub\Hosts\Microsoft.ServiceHub.Host.CLR\vs_installerservice.exe"
 	if ((Test-ExecutableFile $Program) -or $ForceLoad)
 	{