+++ Rules/IPv4/Outbound/Software/Microsoft/MicrosoftOffice.ps1
@@ -223,7 +223,7 @@ to give you a way to see the state of files you are uploading to a SharePoint se
 		New-NetFirewallRule -Platform $Platform `
 			-DisplayName "Outlook (SMTP)" -Service Any -Program $Program `
 			-PolicyStore $PolicyStore -Enabled False -Action Allow -Group $Group -Profile $DefaultProfile -InterfaceType $DefaultInterface `
-			-Direction $Direction -Protocol TCP -LocalAddress Any -RemoteAddress Internet4 -LocalPort Any -RemotePort 25 `
+			-Direction $Direction -Protocol TCP -LocalAddress Any -RemoteAddress Internet4 -LocalPort Any -RemotePort 587 `
 			-LocalUser $UsersGroupSDDL `
 			-Description "Outgoing mail server." | Format-RuleOutput
 	}